Move `TransportProtocol` to Talpid package
