+++ android/src/main/kotlin/net/mullvad/mullvadvpn/model/Endpoint.kt
@@ -1,5 +1,6 @@
 package net.mullvad.mullvadvpn.model
 
 import java.net.InetSocketAddress
+import net.mullvad.talpid.net.TransportProtocol
 
 data class Endpoint(val address: InetSocketAddress, val protocol: TransportProtocol)
