+++ mullvad-jni/src/lib.rs
@@ -62,8 +62,6 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "net/mullvad/mullvadvpn/model/RelaySettingsUpdate$CustomTunnelEndpoint",
     "net/mullvad/mullvadvpn/model/RelaySettingsUpdate$RelayConstraintsUpdate",
     "net/mullvad/mullvadvpn/model/Settings",
-    "net/mullvad/mullvadvpn/model/TransportProtocol$Tcp",
-    "net/mullvad/mullvadvpn/model/TransportProtocol$Udp",
     "net/mullvad/mullvadvpn/model/TunConfig",
     "net/mullvad/mullvadvpn/model/TunnelEndpoint",
     "net/mullvad/mullvadvpn/model/TunnelState$Blocked",
@@ -73,6 +71,8 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "net/mullvad/mullvadvpn/model/TunnelState$Disconnecting",
     "net/mullvad/mullvadvpn/MullvadDaemon",
     "net/mullvad/mullvadvpn/MullvadVpnService",
+    "net/mullvad/talpid/net/TransportProtocol$Tcp",
+    "net/mullvad/talpid/net/TransportProtocol$Udp",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Block",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Nothing",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Reconnect",
